--- app/src/androidTest/java/chat/rocket/android/UITestConfig.kt
@@ -1,18 +1,0 @@
-package chat.rocket.android
-
-class UITestConfig {
-    companion object {
-        const val SERVER: String = "open.rocket.chat"
-        const val serverUrl: String = "serverUrl"
-        const val NAME: String = "user"
-        const val EMAIL: String = "abc@gmail.com"
-
-        //Organisation using RC forks should create a user with below details before running tests
-        const val USERNAME: String = "user121"
-        const val PASSWORD: String = "123456"
-
-        const val CODE = "1234"
-        const val USER_ID = "user_id"
-        const val AUTH_TOKEN = "auth_token"
-    }
-}
