+++ app/src/androidTest/java/chat/rocket/android/authentication/resetpassword/ui/ResetPasswordFragmentTest.kt
@@ -6,7 +6,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
-import chat.rocket.android.UITestConfig.Companion.EMAIL
+import testConfig.Config.Companion.EMAIL
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.matchers.withHint
