+++ app/src/androidTest/java/chat/rocket/android/authentication/server/ui/ServerFragmentTest.kt
@@ -6,7 +6,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.filters.LargeTest
 import androidx.test.rule.ActivityTestRule
-import chat.rocket.android.UITestConfig.Companion.SERVER
+import testConfig.Config.Companion.SERVER
 import chat.rocket.android.R
 import chat.rocket.android.analytics.event.ScreenViewEvent
 import chat.rocket.android.authentication.ui.AuthenticationActivity
