+++ app/src/main/java/chat/rocket/android/authentication/registerusername/presentation/RegisterUsernamePresenter.kt
@@ -3,7 +3,6 @@ package chat.rocket.android.authentication.registerusername.presentation
 import chat.rocket.android.analytics.AnalyticsManager
 import chat.rocket.android.analytics.event.AuthenticationEvent
 import chat.rocket.android.authentication.presentation.AuthenticationNavigator
-import chat.rocket.android.authentication.presentation.AuthenticationPresenter.Companion.defaultServer
 import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.server.domain.GetConnectingServerInteractor
 import chat.rocket.android.server.domain.GetSettingsInteractor
@@ -25,6 +24,7 @@ import chat.rocket.common.model.Token
 import chat.rocket.common.util.ifNull
 import chat.rocket.core.RocketChatClient
 import chat.rocket.core.internal.rest.updateOwnBasicInformation
+import testConfig.Config.Companion.defaultTestServer
 import javax.inject.Inject
 
 class RegisterUsernamePresenter @Inject constructor(
@@ -39,7 +39,7 @@ class RegisterUsernamePresenter @Inject constructor(
     val factory: RocketChatClientFactory,
     val settingsInteractor: GetSettingsInteractor
 ) {
-    private val currentServer = serverInteractor.get()?: defaultServer
+    private val currentServer = serverInteractor.get()?: defaultTestServer
     private val client: RocketChatClient = factory.get(currentServer)
     private var settings: PublicSettings = settingsInteractor.get(currentServer)
     private val token = tokenRepository.get(currentServer)
