+++ app/src/test/java/chat/rocket/android/authentication/login/presentation/LoginPresenterTest.kt
@@ -1,6 +1,5 @@
 package chat.rocket.android.authentication.login.presentation
 
-import chat.rocket.android.UnitTestConfig.Companion.currentServer
 import chat.rocket.android.analytics.AnalyticsManager
 import chat.rocket.android.authentication.presentation.AuthenticationNavigator
 import chat.rocket.android.core.lifecycle.CancelStrategy
@@ -13,6 +12,7 @@ import org.junit.Before
 import org.junit.Test
 import org.mockito.Mockito.*
 import org.mockito.MockitoAnnotations
+import testConfig.Config.Companion.currentServer
 
 class LoginPresenterTest {
