+++ app/src/test/java/chat/rocket/android/authentication/loginoptions/presentation/LoginOptionsPresenterTest.kt
@@ -1,6 +1,5 @@
 package chat.rocket.android.authentication.loginoptions.presentation
 
-import chat.rocket.android.UnitTestConfig.Companion.currentServer
 import chat.rocket.android.analytics.AnalyticsManager
 import chat.rocket.android.authentication.presentation.AuthenticationNavigator
 import chat.rocket.android.core.lifecycle.CancelStrategy
@@ -13,6 +12,7 @@ import org.mockito.Mockito
 import org.mockito.Mockito.`when`
 import org.mockito.Mockito.verify
 import org.mockito.MockitoAnnotations
+import testConfig.Config.Companion.currentServer
 
 
 class LoginOptionsPresenterTest {
