+++ app/src/test/java/chat/rocket/android/authentication/onboarding/presentation/OnBoardingPresenterTest.kt
@@ -1,6 +1,5 @@
 package chat.rocket.android.authentication.onboarding.presentation
 
-import chat.rocket.android.UnitTestConfig.Companion.communityServerUrl
 import chat.rocket.android.authentication.presentation.AuthenticationNavigator
 import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.server.domain.GetAccountsInteractor
@@ -13,6 +12,7 @@ import org.junit.Test
 import org.mockito.Mockito
 import org.mockito.Mockito.verify
 import org.mockito.MockitoAnnotations
+import testConfig.Config.Companion.communityServerUrl
 
 
 class OnBoardingPresenterTest {
