+++ app/src/test/java/chat/rocket/android/authentication/presentation/AuthenticationPresenterTest.kt
@@ -1,10 +1,5 @@
 package chat.rocket.android.authentication.presentation
 
-import chat.rocket.android.UnitTestConfig.Companion.PRIVACY_POLICY
-import chat.rocket.android.UnitTestConfig.Companion.TERMS_OF_SERVICE
-import chat.rocket.android.UnitTestConfig.Companion.currentServer
-import chat.rocket.android.UnitTestConfig.Companion.privacyPolicyUrl
-import chat.rocket.android.UnitTestConfig.Companion.termsOfServiceUrl
 import chat.rocket.android.core.lifecycle.CancelStrategy
 import chat.rocket.android.infrastructure.LocalRepository
 import chat.rocket.android.server.domain.*
@@ -13,6 +8,11 @@ import org.junit.Test
 import org.mockito.Mockito
 import org.mockito.Mockito.verify
 import org.mockito.MockitoAnnotations
+import testConfig.Config.Companion.PRIVACY_POLICY
+import testConfig.Config.Companion.TERMS_OF_SERVICE
+import testConfig.Config.Companion.currentServer
+import testConfig.Config.Companion.privacyPolicyUrl
+import testConfig.Config.Companion.termsOfServiceUrl
 
 
 class AuthenticationPresenterTest {
